+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Requests\AddUserReq;
 use App\Http\Requests\ProfilePassReq;
 use App\Http\Requests\ProfileReq;
+use App\Models\Log;
 use App\Models\Role;
 use App\User;
 use Illuminate\Http\Request;
@@ -17,11 +18,6 @@ use Intervention\Image\ImageManagerStatic as Image;
 class UserController extends Controller
 {
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         if (Auth::user()->can('viewusers')) {
@@ -32,27 +28,22 @@ class UserController extends Controller
             }
             return view('admin.users.index', compact('users'));
         } else {
+            Log::newLog("Usuário tentou acesso area restrita USERS, user: " . Auth::user()->name);
             return view('admin.layout.403');
         }
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
-        $roles = Role::pluck('name', 'id')->all();
-        return view('admin.users.add', compact('roles'));
+        if (Auth::user()->can('createusers')) {
+            $roles = Role::pluck('name', 'id')->all();
+            return view('admin.users.add', compact('roles'));
+        } else {
+            Log::newLog("Usuário tentou criar area restrita USERS, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \App\Http\Requests\AddUserReq $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(AddUserReq $request)
     {
         $user = new User();
@@ -64,43 +55,27 @@ class UserController extends Controller
 
         $user->attachRole($request->get('role'));
 
+        Log::newLog("Usuário criou nova conta, user: " . Auth::user()->name);
+
         flash("Usuário cadastrado com sucesso!")->success();
         return redirect('dashboard/users');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
-        $user = User::findOrFail($id);
-        $roles = Role::pluck('name', 'id')->all();
-        $userRole = $user->roles->pluck('id', $id)->all();
+        if (Auth::user()->can('editusers')) {
+            $user = User::findOrFail($id);
+            $roles = Role::pluck('name', 'id')->all();
+            $userRole = $user->roles->pluck('id', $id)->all();
 
-        return view('admin.users.edit', compact('user', 'roles', 'userRole'));
+            return view('admin.users.edit', compact('user', 'roles', 'userRole'));
+        } else {
+            Log::newLog("Usuário tentou editar area restrita USERS-$id, user: " . Auth::user()->name);
+            return view('admin.layout.403');
+        }
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
         $user = User::findOrFail($id);
@@ -116,25 +91,28 @@ class UserController extends Controller
 
         $user->attachRole($request->get('role'));
 
+        Log::newLog("Usuário alterou/criou conta, user: " . Auth::user()->name);
+
         flash("Usuário alterado com sucesso!")->success();
         return redirect('dashboard/users');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
-        if ($id == Auth::user()->id) {
-            flash("Você não pode deletar você mesmo")->info();
-            return redirect()->back();
+        if (Auth::user()->can('deleteusers')) {
+            if ($id == Auth::user()->id) {
+                Log::newLog("Usuário tentou deletar propria conta, user: " . Auth::user()->name);
+                flash("Você não pode deletar você mesmo")->info();
+                return redirect()->back();
+            } else {
+                User::findOrFail($id)->delete();
+                Log::newLog("Usuário deletou conta-$id, user: " . Auth::user()->name);
+                flash("Usuário deletado.")->success();
+                return redirect()->back();
+            }
         } else {
-            User::findOrFail($id)->delete();
-            flash("Usuário deletado.")->success();
-            return redirect()->back();
+            Log::newLog("Usuário deletar acesso area restrita USERS-$id, user: " . Auth::user()->name);
+            return view('admin.layout.403');
         }
     }
 
@@ -170,6 +148,8 @@ class UserController extends Controller
 
         $user->save();
 
+        Log::newLog("Usuário alterou sua conta, user: " . Auth::user()->name);
+
         flash("Perfil atualizado com sucesso.")->success();
         return redirect('dashboard/profile');
     }
@@ -179,6 +159,8 @@ class UserController extends Controller
         $user = User::findOrFail(Auth::user()->id);
         $user->password = Hash::make($request->get('password'));
         $user->update();
+
+        Log::newLog("Usuário alterou sua senha, user: " . Auth::user()->name);
 
         flash("Senha atualizada com sucesso!")->success();
         return redirect('dashboard/profile');
